IMU sampling rate and file transfers
5Hz is too low of a samling rate, changed to 10 Hz. Higher is ideal, but the app should collect data for an entire day, so too high of a sampling rate can cause file transfer to server fail due to memory constraints.

Discovered that the higher sampling rates used over the past few days caused .zip files of >10MB to be produced. These could not be sent reliably over the connection. This is fixed by sending only a day's worth of files at a time. I.e. for n days of data, we zip n folders and make n transfers