Fixed timestamps (more consistent) and increase sampling rate (25Hz)
Timestamps should be more consistent now. A reference to system start time is generated when the app opens. This is used with the timestamps generated by sensors to get a more accurate sample time.

Actual sampling time discovered to be around 17-18 Hz, so increase setting to 25 Hz. 20 Hz is the lower limit for activity recognition, according to Jian Wu